Ball is launching
solution seems to have been I had a MainManager and a BoardManager both in the StartMenu scene. Moving the BoardManager to the Main scene with the proper prefabs attached resolved this difficult issue.

Main solution: Keep Instances in the scenes that use them. Create those instances in the proper scene.